# Conflicts:
#	Assets/Scripts/Sound/AudioManager.cs

+++ Assets/Scripts/Sound/AudioManager.cs
@@ -40,7 +40,8 @@ public class AudioManager : MonoBehaviour
         }
     }
 
-    private void OnDangerLevelUpdate(float dangerLevel)
+
+    private void OnDangerLevelUpdate(int dangerLevel)
     {
 
     }
